queryset-refactor: The Oracle changes necessary for [7426]. I can't test these at the moment, but they should be close to correct.
Refs #6956.


git-svn-id: http://code.djangoproject.com/svn/django/branches/queryset-refactor@7427 bcc190cf-cafb-0310-a4f2-bffc1f526a37
